refactor: extract processSale to separate file to satisfy max-lines
